Moved The Issue Types To lib/types File
