+++ frontend/components/IssueInputBox/IssueInputBox.tsx
@@ -13,24 +13,10 @@ import { supabase } from '@/lib/globals';
 import LocationAutocomplete from '@/components/LocationAutocomplete/LocationAutocomplete';
 import Dropdown from "@/components/Dropdown/Dropdown";
 import { Image as LucideImage, X } from 'lucide-react';
-import { LocationType } from '@/lib/types';
+import { LocationType, IssueInputBoxProps } from "@/lib/types";
 import Image from 'next/image';
 
 const MAX_CHAR_COUNT = 500;
-
-interface IssueInputBoxProps {
-  user: {
-    user_id: string;
-    email_address: string;
-    username: string;
-    fullname: string;
-    image_url: string;
-    bio: string;
-    is_owner: boolean;
-    total_issues: number;
-    resolved_issues: number;
-  } | null;
-}
 
 const IssueInputBox: React.FC<IssueInputBoxProps> = ({ user }) => {
   const [content, setContent] = useState('');
